I tried to add the functionality to the game-mode selection menu I implemented earlier. As already discussed, I cannot test it because setting up a local database and all this is not working for me.
To be completely honest, I am not sure if this is the right approach. I do not have a coding background. I tried everything I could over the last couple of weeks, but I think if you want to implement different game modes someone professional shouold start over instead of going with my changes.

It was still fun and a good opportunity to test and refine my skills, thank you